fixtures/python/pycryptodome-legacy/requirements.txt
pycryptodome==3.15.0
requests==2.28.1